removing unnecessary mixin
