+++ test/index.html
@@ -16,7 +16,6 @@
 				'../components/inputs/test/input-checkbox.html',
 				'../components/inputs/test/input-search.html',
 				'../components/inputs/test/input-text.html',
-				'../components/inputs/test/input-text-mixin.html',
 				'../components/link/test/link.html',
 				'../components/menu/test/menu-item.html',
 				'../components/menu/test/menu-item-mixin.html',
